+++ lib/ruby/lib/alces/packager/option_set.rb
@@ -22,7 +22,7 @@
 module Alces
   module Packager
     class OptionSet
-      attr_accessor :compiler, :variant, :depot, :args, :tag, :modules, :binary, :binary_depends, :binary_only
+      attr_accessor :compiler, :variant, :depot, :args, :tag, :modules, :binary, :binary_depends, :binary_only, :defaults
       attr_accessor :global, :latest, :yes, :non_interactive, :verbose
       attr_accessor :ignore_bad, :accept_elf, :accept_bad, :patch_binary, :output, :packages, :notify
       attr_accessor :tree, :ignore_satisfied
@@ -40,6 +40,7 @@ module Alces
           self.binary = options.binary
           self.binary_depends = options.binary_depends
           self.binary_only = options.binary_only
+          self.defaults = options.defaults
 
           self.global = options.global
           self.latest = options.latest